+++ src/ViceCode.Analyzers.Test/Helpers/DiagnosticResult.cs
