+++ src/ViceCode.Analyzers.Test/MakeConstUnitTests.cs
