+++ src/ViceCode.Analyzers.Test/ViceCode.Analyzers.Test.csproj
