+++ src/ViceCode.Analyzers.Vsix/ViceCode.Analyzers.Vsix.csproj
