+++ src/ViceCode.Analyzers.Vsix/source.extension.vsixmanifest
