+++ src/ViceCode.Analyzers/Resources.resx
