+++ src/ViceCode.Analyzers/Utils/Helper.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace ViceCode.Analyzers
+namespace ViceCode.Analyzers.Utils
 {
     internal static class Helper
     {
