+++ src/ViceCode.Analyzers/ViceCode.Analyzers.csproj
