+++ src/ViceCode.Analyzers/tools/install.ps1
