+++ src/ViceCode.Analyzers/tools/uninstall.ps1
